v1.0.0-alpha.4 - Actions routing
