+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "hyperspan",
-  "version": "1.0.0-alpha.3",
+  "version": "1.0.0-alpha.4",
   "description": "Hyperspan CLI - for @hyperspan/framework",
   "type": "module",
   "public": true,
